open-pi-zero/pyproject.toml
[project]
name = "open-pi-zero"
version = "0.1.1"
description = "Re-implementation of Pi0 vision-language-action (VLA) model from Physical Intelligence"
authors = [
  {name = "Allen Z. Ren", email = "allenzren1@gmail.com"},
]
readme = "README.md"
requires-python = "==3.10.*"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "bitsandbytes==0.45.0",
    "einops",
    "gsutil>=5.32",
    "hydra-core",
    "imageio",
    "matplotlib",
    "numpy==1.26.4",
    "omegaconf",
    "pillow",
    "pre-commit>=4.0.1",
    "pretty_errors",
    "protobuf==3.20.3",
    "tensorflow==2.15.0",
    "tensorflow_datasets==4.9.2",
    "torch==2.5.1", # see https://github.com/pytorch/pytorch/pull/138587
    "transformers==4.47.1",
    "tqdm",
    "wandb",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = []

[tool.ruff]
line-length = 88
target-version = "py310"
extend-exclude = ["src/data/obs_transforms.py", "src/data/utils/data_utils.py"]

[tool.ruff.lint]
select = ["A", "B", "E", "F", "I", "RUF", "W"]
ignore = ["E203", "E501", "B006", "B026", "B905"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403"]
